refactor: lift grok/codex session paths into @genesiscz/utils
Worker session paths and parseTurnLog were already env-only but lived
under src/grok and src/codex, so utils transcript code could not import
them. Native CLI roots (claude/grok/codex) were duplicated in ai-spend
drivers.

Move the path helpers and streaming-json parser into src/utils, leave
re-export shims at the old paths, and point the spend drivers at
nativeSessionRoots.
